+++ common/src/terrain/site.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq)]
+pub enum SitesKind {
+    Dungeon,
+    Cave,
+    Settlement,
+    Castle,
+    Void,
+}