+++ voxygen/src/main.rs
@@ -4,12 +4,12 @@
 #![recursion_limit = "2048"]
 
 use veloren_voxygen::{
-    audio::{self, AudioFrontend},
+    audio::AudioFrontend,
     i18n::{self, i18n_asset_key, Localization},
     logging,
     profile::Profile,
     run,
-    settings::{AudioOutput, Settings},
+    settings::Settings,
     window::Window,
     GlobalState,
 };
@@ -142,13 +142,14 @@ fn main() {
     anim::init();
 
     // Setup audio
-    let mut audio = match settings.audio.output {
-        AudioOutput::Off => None,
-        AudioOutput::Automatic => audio::get_default_device(),
-        AudioOutput::Device(ref dev) => Some(dev.clone()),
-    }
-    .map(|dev| AudioFrontend::new(dev, settings.audio.max_sfx_channels))
-    .unwrap_or_else(AudioFrontend::no_audio);
+    //let mut audio = match settings.audio.output {
+    //    AudioOutput::Off => None,
+    //    AudioOutput::Automatic => audio::get_default_device(),
+    //    AudioOutput::Device(ref dev) => Some(dev.clone()),
+    //}
+    //.map(|dev| AudioFrontend::new(dev, settings.audio.max_sfx_channels))
+    //.unwrap_or_else(AudioFrontend::no_audio);
+    let mut audio = AudioFrontend::new(settings.audio.max_sfx_channels);
 
     audio.set_music_volume(settings.audio.music_volume);
     audio.set_sfx_volume(settings.audio.sfx_volume);